Fix a case with enter inside curly brackets.
This is a case that never happens in c, but I was writing some js and had a =>{}); situation, and when I would hit enter inside the curly brackets, the paren and semicolon would stay on the line it was on. I believe this solution works, but I have done limited testing, so feel free to test further.